# backend/openapi.yaml
openapi: 3.1.0
info:
  title: ETMS API
  version: "1.0.0"
  description: >
    Employee Transportation Management System REST API. Human-readable contract
    in docs/etms/06-api-spec.md; this file covers the endpoints implemented in the
    backend service. All /v1 routes require a Bearer JWT carrying tenant + permissions.
servers:
  - url: https://api.{region}.etms.app/v1
    variables: { region: { default: dev } }
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        request_id: { type: string }
...
security:
  - bearerAuth: []
paths:
...
  /v1/auth/login:
    post:
      security: []
      summary: Verify credentials → JWT carrying tenant + effective permissions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantSlug, email, password]
              properties:
                tenantSlug: { type: string }
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200": { description: Access token + refresh token + user }
        "401": { description: Invalid credentials, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
        "429": { description: Rate limited (per-IP) or account locked (per-account, progressive) — see Retry-After }
  /v1/auth/password:
    post:
      summary: Change password (verifies current, enforces strength, revokes other sessions)
      responses:
        "200": { description: Changed }
        "401": { description: Current password incorrect }
        "422": { description: New password too weak }
  /v1/auth/refresh:
    post:
      security: []
      summary: Rotate a refresh token → new access + refresh (old token single-use)
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [refreshToken], properties: { refreshToken: { type: string } } }
      responses:
        "200": { description: New tokens }
        "401": { description: Invalid/expired/revoked refresh token }
  /v1/auth/logout:
    post:
      security: []
      summary: Revoke a refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [refreshToken], properties: { refreshToken: { type: string } } }
      responses: { "200": { description: Revoked } }
...
  /v1/auth/mfa/verify:
    post:
      summary: Confirm TOTP enrollment with a code
      responses: { "200": { description: MFA enabled }, "401": { description: Invalid code } }
  /v1/auth/mfa/challenge:
    post:
      summary: Step-up — verify a code → short-lived token with mfa:true
      responses: { "200": { description: Step-up token }, "401": { description: Invalid code } }
...